updated routes as mentioned
