+++ app.js
@@ -1,10 +1,9 @@
 const express = require('express');
 const app = express();
 const bodyParser = require('body-parser');
-const routes = require("./routes/neonRoute");
+const routes = require("./routes/routes");
 
-//path to neon route
-const neonRoute = require('./routes/neonRoute');
+
 
 
 app.use(bodyParser.json());
@@ -13,14 +12,14 @@ app.use(bodyParser.urlencoded({
 }));
 
 
-//route for checking whether number is neon or not
-//the complete route is /check/neon/:num
-app.use("/check", neonRoute);
+
+
 app.use(routes);
 
 app.get('/', (req, res) => {
     res.send('Backend running successfully');
 })
+
 
 
 //app is running here
